app_Bankruptcy/requirements.txt
pandas==1.3.4
Pillow==8.4.0
scikit_learn==1.0.1
streamlit<2.0
